third_party/WebKit/LayoutTests/fast/canvas/font-no-zoom.html

<!DOCTYPE html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<style>
canvas {
  font-family: sans-serif;
  font-size: 100px;
  zoom: 3;
}
</style>
<canvas width="100" height="100"></canvas>
<script>
test(function() {
  var ctx = document.querySelector('canvas').getContext('2d');
  // Set same font as will be computed on the canvas element.
  ctx.font = "100px sans-serif";
  assert_equals(ctx.font, "100px sans-serif");
}, "Zoom does not affect the font size.");
</script>
